disclosure_system/requirements.txt
# Phase 1 완료 - 기본 요소
# (표준 라이브러리만 사용)

# Phase 2 추가 사항:
streamlit>=1.28.0          # 웹 인터페이스
pandas>=1.5.0              # 데이터 처리
openpyxl>=3.1.2           # Excel 파일 지원

pymupdf>=1.20.0            # PDF 파싱
google-generativeai>=0.3.0 # Google Gemini AI
python-dotenv>=0.20.0      # 환경변수 관리
matplotlib>=3.5.0          # 차트 시각화
plotly>=5.10.0             # 인터랙티브 시각화
requests>=2.28.0           # 국가법령정보 API 호출
